+++ packages/ens-controller/src/EnsController.test.ts
@@ -56,8 +56,10 @@ const name = 'EnsController';
  * @returns A restricted controller messenger.
  */
 function getMessenger() {
-  return new ControllerMessenger().getRestricted<typeof name, never, never>({
+  return new ControllerMessenger().getRestricted({
     name,
+    allowedActions: [],
+    allowedEvents: [],
   });
 }