+++ packages/permission-controller/src/SubjectMetadataController.test.ts
@@ -5,7 +5,6 @@ import type { HasPermissions } from './PermissionController';
 import type {
   SubjectMetadataControllerActions,
   SubjectMetadataControllerEvents,
-  SubjectMetadataControllerMessenger,
 } from './SubjectMetadataController';
 import {
   SubjectMetadataController,
@@ -34,12 +33,12 @@ function getSubjectMetadataControllerMessenger() {
   return [
     controllerMessenger.getRestricted<
       typeof controllerName,
-      HasPermissions['type'],
-      never
+      HasPermissions['type']
     >({
       name: controllerName,
       allowedActions: ['PermissionController:hasPermissions'],
-    }) as SubjectMetadataControllerMessenger,
+      allowedEvents: [],
+    }),
     hasPermissionsSpy,
   ] as const;
 }