+++ packages/permission-log-controller/tests/PermissionLogController.test.ts
@@ -41,12 +41,10 @@ const initController = ({
   restrictedMethods: Set<string>;
   state?: Partial<PermissionLogControllerState>;
 }): PermissionLogController => {
-  const messenger = new ControllerMessenger().getRestricted<
-    typeof name,
-    never,
-    never
-  >({
+  const messenger = new ControllerMessenger().getRestricted({
     name,
+    allowedActions: [],
+    allowedEvents: [],
   });
   return new PermissionLogController({
     messenger,