+++ packages/transaction-controller/src/TransactionController.ts
@@ -84,6 +84,7 @@ import {
   getAndFormatTransactionsForNonceTracker,
   getNextNonce,
 } from './utils/nonce';
+import { getSimulationData } from './utils/simulation';
 import {
   updatePostTransactionBalance,
   updateSwapsTransaction,
@@ -245,6 +246,7 @@ export type PendingTransactionOptions = {
  * @property getSelectedAddress - Gets the address of the currently selected account.
  * @property incomingTransactions - Configuration options for incoming transaction support.
  * @property isMultichainEnabled - Enable multichain support.
+ * @property isSimulationEnabled - Whether new transactions will be automatically simulated.
  * @property messenger - The controller messenger.
  * @property onNetworkStateChange - Allows subscribing to network controller state changes.
  * @property pendingTransactions - Configuration options for pending transaction support.
@@ -280,6 +282,7 @@ export type TransactionControllerOptions = {
   getSelectedAddress: () => string;
   incomingTransactions?: IncomingTransactionOptions;
   isMultichainEnabled: boolean;
+  isSimulationEnabled?: () => boolean;
   messenger: TransactionControllerMessenger;
   onNetworkStateChange: (listener: (state: NetworkState) => void) => void;
   pendingTransactions?: PendingTransactionOptions;
@@ -602,6 +605,8 @@ export class TransactionController extends BaseController<
 
   #transactionHistoryLimit: number;
 
+  #isSimulationEnabled: () => boolean;
+
   private readonly afterSign: (
     transactionMeta: TransactionMeta,
     signedTx: TypedTransaction,
@@ -697,6 +702,7 @@ export class TransactionController extends BaseController<
    * @param options.getSelectedAddress - Gets the address of the currently selected account.
    * @param options.incomingTransactions - Configuration options for incoming transaction support.
    * @param options.isMultichainEnabled - Enable multichain support.
+   * @param options.isSimulationEnabled - Whether new transactions will be automatically simulated.
    * @param options.messenger - The controller messenger.
    * @param options.onNetworkStateChange - Allows subscribing to network controller state changes.
    * @param options.pendingTransactions - Configuration options for pending transaction support.
@@ -723,6 +729,7 @@ export class TransactionController extends BaseController<
     getSelectedAddress,
     incomingTransactions = {},
     isMultichainEnabled = false,
+    isSimulationEnabled,
     messenger,
     onNetworkStateChange,
     pendingTransactions = {},
@@ -748,6 +755,7 @@ export class TransactionController extends BaseController<
     this.isSendFlowHistoryDisabled = disableSendFlowHistory ?? false;
     this.isHistoryDisabled = disableHistory ?? false;
     this.isSwapsDisabled = disableSwaps ?? false;
+    this.#isSimulationEnabled = isSimulationEnabled ?? (() => true);
     // @ts-expect-error the type in eth-method-registry is inappropriate and should be changed
     this.registry = new MethodRegistry({ provider });
     this.getSavedGasFees = getSavedGasFees ?? ((_chainId) => undefined);
@@ -1024,7 +1032,10 @@ export class TransactionController extends BaseController<
           networkClientId,
         };
 
-    await this.updateGasProperties(addedTransactionMeta);
+    await Promise.all([
+      this.updateGasProperties(addedTransactionMeta),
+      this.#simulateTransaction(addedTransactionMeta),
+    ]);
 
     // Checks if a transaction already exists with a given actionId
     if (!existingTransactionMeta) {
@@ -3473,5 +3484,25 @@ export class TransactionController extends BaseController<
       );
       state.transactions[index] = transactionWithUpdatedHistory;
     });
+  }
+
+  async #simulateTransaction(transactionMeta: TransactionMeta) {
+    if (!this.#isSimulationEnabled()) {
+      log('Skipping simulation as disabled');
+      return;
+    }
+
+    const { chainId, txParams } = transactionMeta;
+    const { from, to, value, data } = txParams;
+
+    transactionMeta.simulationData = await getSimulationData({
+      chainId,
+      from: from as Hex,
+      to: to as Hex,
+      value: value as Hex,
+      data: data as Hex,
+    });
+
+    log('Retrieved simulation data', transactionMeta.simulationData);
   }
 }